# src/routes/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Record_string.unknown_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
...
			"RecordId": {
				"type": "string",
				"example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
			},
...
			"PlaintextMessage": {
				"properties": {
					"@type": {
						"type": "string"
					},
					"@id": {
						"type": "string"
					},
					"~thread": {
						"properties": {
							"pthid": {
								"type": "string"
							},
							"thid": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"@type",
					"@id"
				],
				"type": "object",
				"additionalProperties": {}
			},
...
			"AgentMessage": {
				"$ref": "#/components/schemas/PlaintextMessage"
			},
...
			"Routing": {
				"properties": {
					"endpoints": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"recipientKey": {
						"$ref": "#/components/schemas/Key"
					},
					"routingKeys": {
						"items": {
							"$ref": "#/components/schemas/Key"
						},
						"type": "array"
					},
					"mediatorId": {
						"type": "string"
					}
				},
				"required": [
					"endpoints",
					"recipientKey",
					"routingKeys"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateInvitationOptions": {
				"properties": {
					"label": {
						"type": "string"
					},
					"alias": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"goalCode": {
						"type": "string"
					},
					"goal": {
						"type": "string"
					},
					"handshake": {
						"type": "boolean"
					},
					"handshakeProtocols": {
						"items": {
							"$ref": "#/components/schemas/HandshakeProtocol"
						},
						"type": "array"
					},
					"messages": {
						"items": {
							"$ref": "#/components/schemas/AgentMessage"
						},
						"type": "array"
					},
					"multiUseInvitation": {
						"type": "boolean"
					},
					"autoAcceptConnection": {
						"type": "boolean"
					},
					"routing": {
						"$ref": "#/components/schemas/Routing"
					},
					"appendedAttachments": {
						"items": {
							"$ref": "#/components/schemas/Attachment"
						},
						"type": "array"
					},
					"invitationDid": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"RecipientKeyOption": {
				"properties": {
					"recipientKey": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_CreateLegacyInvitationConfig.routing_": {
				"$ref": "#/components/schemas/Pick_CreateLegacyInvitationConfig.Exclude_keyofCreateLegacyInvitationConfig.routing__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"OutOfBandRecord": {
				"$ref": "#/components/schemas/Record_string.unknown_"
			},
...
			"ReceiveInvitationProps": {
				"properties": {
					"alias": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"imageUrl": {
						"type": "string"
					},
					"autoAcceptConnection": {
						"type": "boolean"
					},
					"autoAcceptInvitation": {
						"type": "boolean"
					},
					"reuseConnection": {
						"type": "boolean"
					},
					"acceptInvitationTimeoutMs": {
						"type": "number",
						"format": "double"
					},
					"ourDid": {
						"type": "string"
					},
					"invitation": {
						"$ref": "#/components/schemas/OutOfBandInvitationSchema"
					}
				},
				"required": [
					"invitation"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"apiKey": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "credo-controller",
		"version": "2.0.0",
		"description": "Rest endpoint wrapper for using your agent over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/didcomm/oob": {
			"get": {
				"operationId": "GetAllOutOfBandRecords",
				"responses": {
					"200": {
						"description": "OutOfBandRecord[]",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Record_string.unknown_"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"_tags": {
													"invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
													"state": "await-response",
													"role": "sender",
													"recipientKeyFingerprints": [
														"z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
													]
												},
												"outOfBandInvitation": {
													"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
													"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
													"label": "Aries Test Agent",
													"accept": [
														"didcomm/aip1",
														"didcomm/aip2;env=rfc19"
													],
													"handshake_protocols": [
														"https://didcomm.org/didexchange/1.0",
														"https://didcomm.org/connections/1.0"
													],
													"services": [
														{
															"id": "#inline-0",
															"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
															"type": "did-communication",
															"recipientKeys": [
																"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
															],
															"routingKeys": []
														}
													]
												},
												"metadata": {},
												"id": "42a95528-0e30-4f86-a462-0efb02178b53",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"reusable": false
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Retrieve all out of band records",
				"tags": [
					"DIDComm - Out Of Band"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"description": "invitation identifier",
						"in": "query",
						"name": "invitationId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
		"/didcomm/oob/{outOfBandId}": {
			"get": {
				"operationId": "GetOutOfBandRecordById",
				"responses": {
					"200": {
						"description": "OutOfBandRecord",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Record_string.unknown_"
								},
								"examples": {
									"Example 1": {
										"value": {
											"_tags": {
												"invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
												"state": "await-response",
												"role": "sender",
												"recipientKeyFingerprints": [
													"z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
												]
											},
											"outOfBandInvitation": {
												"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
												"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
												"label": "Aries Test Agent",
												"accept": [
													"didcomm/aip1",
													"didcomm/aip2;env=rfc19"
												],
												"handshake_protocols": [
													"https://didcomm.org/didexchange/1.0",
													"https://didcomm.org/connections/1.0"
												],
												"services": [
													{
														"id": "#inline-0",
														"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
														"type": "did-communication",
														"recipientKeys": [
															"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
														],
														"routingKeys": []
													}
												]
											},
											"metadata": {},
											"id": "42a95528-0e30-4f86-a462-0efb02178b53",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"reusable": false
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieve an out of band record by id",
				"tags": [
					"DIDComm - Out Of Band"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "outOfBandId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteOutOfBandRecord",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Deletes an out of band record from the repository.",
				"tags": [
					"DIDComm - Out Of Band"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [
					{
						"description": "Record identifier",
						"in": "path",
						"name": "outOfBandId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
		"/didcomm/oob/create-invitation": {
			"post": {
				"operationId": "CreateInvitation",
				"responses": {
					"200": {
						"description": "Out of band record",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"invitationDid": {
											"type": "string"
										},
										"outOfBandRecord": {
											"$ref": "#/components/schemas/Record_string.unknown_"
										},
										"invitation": {
											"$ref": "#/components/schemas/PlaintextMessage"
										},
										"invitationUrl": {
											"type": "string"
										}
									},
									"required": [
										"invitationDid",
										"outOfBandRecord",
										"invitation",
										"invitationUrl"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"invitationUrl": "string",
											"invitation": {
												"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
												"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
												"label": "Aries Test Agent",
												"accept": [
													"didcomm/aip1",
													"didcomm/aip2;env=rfc19"
												],
												"handshake_protocols": [
													"https://didcomm.org/didexchange/1.0",
													"https://didcomm.org/connections/1.0"
												],
												"services": [
													{
														"id": "#inline-0",
														"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
														"type": "did-communication",
														"recipientKeys": [
															"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
														],
														"routingKeys": []
													}
												]
											},
											"outOfBandRecord": {
												"_tags": {
													"invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
													"state": "await-response",
													"role": "sender",
													"recipientKeyFingerprints": [
														"z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
													]
												},
												"outOfBandInvitation": {
													"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
													"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
													"label": "Aries Test Agent",
													"accept": [
														"didcomm/aip1",
														"didcomm/aip2;env=rfc19"
													],
													"handshake_protocols": [
														"https://didcomm.org/didexchange/1.0",
														"https://didcomm.org/connections/1.0"
													],
													"services": [
														{
															"id": "#inline-0",
															"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
															"type": "did-communication",
															"recipientKeys": [
																"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
															],
															"routingKeys": []
														}
													]
												},
												"metadata": {},
												"id": "42a95528-0e30-4f86-a462-0efb02178b53",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"reusable": false
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates an outbound out-of-band record containing out-of-band invitation message defined in\nAries RFC 0434: Out-of-Band Protocol 1.1.",
				"tags": [
					"DIDComm - Out Of Band"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "configuration of how out-of-band invitation should be created",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/CreateInvitationOptions"
									},
									{
										"$ref": "#/components/schemas/RecipientKeyOption"
									}
								],
								"description": "configuration of how out-of-band invitation should be created"
							}
						}
					}
				}
			}
		},
		"/didcomm/oob/create-legacy-invitation": {
			"post": {
				"operationId": "CreateLegacyInvitation",
				"responses": {
					"200": {
						"description": "out-of-band record and invitation",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"outOfBandRecord": {
											"$ref": "#/components/schemas/Record_string.unknown_"
										},
										"invitation": {
											"$ref": "#/components/schemas/PlaintextMessage"
										},
										"invitationUrl": {
											"type": "string"
										},
										"recipientKey": {
											"type": "string"
										}
									},
									"required": [
										"outOfBandRecord",
										"invitation",
										"invitationUrl"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"invitation": {
												"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
												"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
												"label": "Aries Test Agent",
												"accept": [
													"didcomm/aip1",
													"didcomm/aip2;env=rfc19"
												],
												"handshake_protocols": [
													"https://didcomm.org/didexchange/1.0",
													"https://didcomm.org/connections/1.0"
												],
												"services": [
													{
														"id": "#inline-0",
														"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
														"type": "did-communication",
														"recipientKeys": [
															"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
														],
														"routingKeys": []
													}
												]
											},
											"outOfBandRecord": {
												"_tags": {
													"invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
													"state": "await-response",
													"role": "sender",
													"recipientKeyFingerprints": [
														"z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
													]
												},
												"outOfBandInvitation": {
													"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
													"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
													"label": "Aries Test Agent",
													"accept": [
														"didcomm/aip1",
														"didcomm/aip2;env=rfc19"
													],
													"handshake_protocols": [
														"https://didcomm.org/didexchange/1.0",
														"https://didcomm.org/connections/1.0"
													],
													"services": [
														{
															"id": "#inline-0",
															"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
															"type": "did-communication",
															"recipientKeys": [
																"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
															],
															"routingKeys": []
														}
													]
												},
												"metadata": {},
												"id": "42a95528-0e30-4f86-a462-0efb02178b53",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"reusable": false
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates an outbound out-of-band record in the same way how `createInvitation` method does it,\nbut it also converts out-of-band invitation message to an \"legacy\" invitation message defined\nin RFC 0160: Connection Protocol and returns it together with out-of-band record.",
				"tags": [
					"DIDComm - Out Of Band"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "configuration of how a invitation should be created",
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"allOf": [
									{
										"$ref": "#/components/schemas/Omit_CreateLegacyInvitationConfig.routing_"
									},
									{
										"$ref": "#/components/schemas/RecipientKeyOption"
									}
								],
								"description": "configuration of how a invitation should be created"
							}
						}
					}
				}
			}
		},
		"/didcomm/oob/create-legacy-connectionless-invitation": {
			"post": {
				"operationId": "CreateLegacyConnectionlessInvitation",
				"responses": {
					"200": {
						"description": "a message and a invitationUrl",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"outOfBandRecord": {
											"$ref": "#/components/schemas/OutOfBandRecord"
										},
										"invitationUrl": {
											"type": "string"
										},
										"message": {
											"$ref": "#/components/schemas/AgentMessage"
										}
									},
									"required": [
										"outOfBandRecord",
										"invitationUrl",
										"message"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": {
												"@id": "eac4ff4e-b4fb-4c1d-aef3-b29c89d1cc00",
												"@type": "https://didcomm.org/connections/1.0/invitation"
											},
											"invitationUrl": "http://example.com/invitation_url"
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates a new connectionless legacy invitation.",
				"tags": [
					"DIDComm - Out Of Band"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "configuration of how a connection invitation should be created",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"routing": {
										"$ref": "#/components/schemas/Routing"
									},
									"domain": {
										"type": "string"
									},
									"message": {
										"$ref": "#/components/schemas/Record_string.unknown_"
									},
									"recordId": {
										"type": "string"
									}
								},
								"required": [
									"domain",
									"message",
									"recordId"
								],
								"type": "object",
								"description": "configuration of how a connection invitation should be created"
							}
						}
					}
				}
			}
		},
		"/didcomm/oob/receive-invitation": {
			"post": {
				"operationId": "ReceiveInvitation",
				"responses": {
					"200": {
						"description": "out-of-band record and connection record if one has been created.",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"connectionRecord": {
											"$ref": "#/components/schemas/Record_string.unknown_"
										},
										"outOfBandRecord": {
											"$ref": "#/components/schemas/Record_string.unknown_"
										}
									},
									"required": [
										"connectionRecord",
										"outOfBandRecord"
									],
									"type": "object"
								},
								"examples": {
									"Example 1": {
										"value": {
											"outOfBandRecord": {
												"_tags": {
													"invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
													"state": "await-response",
													"role": "sender",
													"recipientKeyFingerprints": [
														"z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
													]
												},
												"outOfBandInvitation": {
													"@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
													"@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
													"label": "Aries Test Agent",
													"accept": [
														"didcomm/aip1",
														"didcomm/aip2;env=rfc19"
													],
													"handshake_protocols": [
														"https://didcomm.org/didexchange/1.0",
														"https://didcomm.org/connections/1.0"
													],
													"services": [
														{
															"id": "#inline-0",
															"serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
															"type": "did-communication",
															"recipientKeys": [
																"did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
															],
															"routingKeys": []
														}
													]
												},
												"metadata": {},
												"id": "42a95528-0e30-4f86-a462-0efb02178b53",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"reusable": false
											},
											"connectionRecord": {
												"_tags": {
													"invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
													"did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
													"invitationKey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz",
													"outOfBandId": "edbc89fe-785f-4774-a288-46012486881d",
													"verkey": "9HG4rJFpLiWf56MWxHj9rgdpErFzim2zEpHuxy1dw7oz"
												},
												"metadata": {},
												"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
												"invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
												"outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
											}
										}
									}
								}
							}
						}
					}
				},
				"description": "Creates inbound out-of-band record and assigns out-of-band invitation message to it if the\nmessage is valid.",
				"tags": [
					"DIDComm - Out Of Band"
				],
				"security": [
					{
						"jwt": [
							"tenant",
							"dedicated"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReceiveInvitationProps"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}